Re-organize sample implementations
The Service sample implementations are re-organized into two categories.

- Protocol Samples: These would be the samples that are used to verify
core protocol changes. The samples are created for demonstrating sample
messages exchanges in each of the FIDOIOT protocol levels. These
samples use hard-coded configuration parameters.

- Component Samples: These samples implement different FIDOIOT
components; Manufacturer service, Reseller service, Rendezvous service,
Owner service, TO0 client and Device client.

Few additional changes:
1. The package name for the samples are also modified from web to samples.
2. Fixed a wrong entry-point reference in http-dispatcher.
