--- service/http-client-samples/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <artifactId>service</artifactId>
-    <groupId>org.fido.iot</groupId>
-    <version>0.3-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>http-client-samples</artifactId>
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>http-client-di-sample</module>
-    <module>http-client-to0-sample</module>
-    <module>http-client-to1-sample</module>
-    <module>http-client-to2-sample</module>
-  </modules>
-
-</project>
