--- service/http-server-samples/http-server-di-sample/src/main/java/org/fido/iot/web/DiContextListener.java
@@ -1,155 +1,0 @@
-// Copyright 2020 Intel Corporation
-// SPDX-License-Identifier: Apache 2.0
-
-package org.fido.iot.web;
-
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.security.cert.Certificate;
-import java.util.List;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
-import javax.sql.DataSource;
-import org.apache.commons.dbcp2.BasicDataSource;
-import org.fido.iot.certutils.PemLoader;
-import org.fido.iot.protocol.CloseableKey;
-import org.fido.iot.protocol.Composite;
-import org.fido.iot.protocol.Const;
-import org.fido.iot.protocol.CryptoService;
-import org.fido.iot.protocol.DiServerService;
-import org.fido.iot.protocol.DiServerStorage;
-import org.fido.iot.protocol.MessageDispatcher;
-import org.fido.iot.protocol.MessagingService;
-import org.fido.iot.storage.CertificateResolver;
-import org.fido.iot.storage.DiDbManager;
-import org.fido.iot.storage.DiDbStorage;
-
-/**
- * Device Initialization servlet Context Listener.
- */
-public class DiContextListener implements ServletContextListener {
-
-  private static final String mfgKeyPem = "-----BEGIN CERTIFICATE-----\n"
-      + "MIIBIjCByaADAgECAgkApNMDrpgPU/EwCgYIKoZIzj0EAwIwDTELMAkGA1UEAwwC\n"
-      + "Q0EwIBcNMTkwNDI0MTQ0NjQ3WhgPMjA1NDA0MTUxNDQ2NDdaMA0xCzAJBgNVBAMM\n"
-      + "AkNBMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAELAJwkDKz/BaWq1Wx7PjkR5W5\n"
-      + "LLIbamgSZeVNUlyFM/t0sMAxAWbvEbDzKu924TX4as3WVjMmfekysx30PlDGJaMQ\n"
-      + "MA4wDAYDVR0TBAUwAwEB/zAKBggqhkjOPQQDAgNIADBFAiEApUGbgjYT0k63AeRA\n"
-      + "tPM2i+VnW6ckYaJyvFLuuWw+QUACIE5w0ntjHLbvwmqgwCfh5T6u8exQdCA2g9Hs\n"
-      + "u53hKcaS\n"
-      + "-----END CERTIFICATE-----\n"
-      + "-----BEGIN EC PARAMETERS-----\n"
-      + "BggqhkjOPQMBBw==\n"
-      + "-----END EC PARAMETERS-----\n"
-      + "-----BEGIN EC PRIVATE KEY-----\n"
-      + "MHcCAQEEIJTKW2/54N85RLJu0C5fEkAwQiKqxRqHzx5PUfd/M66UoAoGCCqGSM49\n"
-      + "AwEHoUQDQgAELAJwkDKz/BaWq1Wx7PjkR5W5LLIbamgSZeVNUlyFM/t0sMAxAWbv\n"
-      + "EbDzKu924TX4as3WVjMmfekysx30PlDGJQ==\n"
-      + "-----END EC PRIVATE KEY-----";
-
-  private final String ownerKeysPem = "-----BEGIN PUBLIC KEY-----\n"
-      + "MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEWVUE2G0GLy8scmAOyQyhcBiF/fSU\n"
-      + "d3i/Og7XDShiJb2IsbCZSRqt1ek15IbeCI5z7BHea2GZGgaK63cyD15gNA==\n"
-      + "-----END PUBLIC KEY-----\n";
-
-  CertificateResolver keyResolver;
-
-  @Override
-  public void contextInitialized(ServletContextEvent sce) {
-    BasicDataSource ds = new BasicDataSource();
-
-    ServletContext sc = sce.getServletContext();
-    ds.setUrl(sc.getInitParameter("db.url"));
-    ds.setDriverClassName("org.h2.Driver");
-    ds.setUsername(sc.getInitParameter("db.user"));
-    ds.setPassword(sc.getInitParameter("db.password"));
-
-    System.out.println(ds.getUrl());
-
-    ds.setMinIdle(5);
-    ds.setMaxIdle(10);
-    ds.setMaxOpenPreparedStatements(100);
-
-    CryptoService cs = new CryptoService();
-
-    sc.setAttribute("datasource", ds);
-    sc.setAttribute("cryptoservice", cs);
-
-    keyResolver = new CertificateResolver() {
-      @Override
-      public CloseableKey getPrivateKey(Certificate cert) {
-
-        return new CloseableKey(
-            PemLoader.loadPrivateKey(mfgKeyPem));
-      }
-
-      @Override
-      public Certificate[] getCertChain(int publicKeyType) {
-        List<Certificate> list = PemLoader.loadCerts(mfgKeyPem);
-        Certificate[] certs = new Certificate[list.size()];
-        list.toArray(certs);
-        return certs;
-      }
-    };
-
-    MessageDispatcher dispatcher = new MessageDispatcher() {
-      @Override
-      protected MessagingService getMessagingService(Composite request) {
-        return createDiService(cs, ds);
-      }
-
-      @Override
-      protected void replied(Composite reply) {
-        sc.log("replied with: " + reply.toString());
-      }
-
-      @Override
-      protected void dispatching(Composite request) {
-        sc.log("dispatching: " + request.toString());
-      }
-
-      @Override
-      protected void failed(Exception e) {
-        StringWriter writer = new StringWriter();
-        try (PrintWriter pw = new PrintWriter(writer)) {
-          e.printStackTrace(pw);
-        }
-        sc.log(writer.toString());
-      }
-    };
-    sc.setAttribute(Const.DISPATCHER_ATTRIBUTE, dispatcher);
-    sc.setAttribute("resolver", keyResolver);
-
-    //create tables
-    DiDbStorage db = new DiDbStorage(cs, ds, keyResolver);
-    DiDbManager manager = new DiDbManager();
-    manager.createTables(ds);
-    manager.addCustomer(ds, 1, "owner", ownerKeysPem);
-    manager.setAutoEnroll(ds, 1);
-
-  }
-
-  @Override
-  public void contextDestroyed(ServletContextEvent sce) {
-  }
-
-  private DiServerService createDiService(CryptoService cs, DataSource ds) {
-    return new DiServerService() {
-      private DiServerStorage storage;
-
-      @Override
-      public DiServerStorage getStorage() {
-        if (storage == null) {
-          storage = new DiDbStorage(cs, ds, keyResolver);
-        }
-        return storage;
-      }
-
-      @Override
-      public CryptoService getCryptoService() {
-        return cs;
-      }
-    };
-  }
-}
