--- service/http-server-samples/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <artifactId>service</artifactId>
-    <groupId>org.fido.iot</groupId>
-    <version>0.3-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>http-server-samples</artifactId>
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>http-server-di-sample</module>
-    <module>http-server-rv-sample</module>
-    <module>http-server-owner-sample</module>
-  </modules>
-
-</project>
