+++ service/protocol-samples/http-client-di-sample/pom.xml
@@ -5,13 +5,13 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <artifactId>http-dispatchers</artifactId>
+    <artifactId>protocol-samples</artifactId>
     <groupId>org.fido.iot</groupId>
     <version>0.3-SNAPSHOT</version>
   </parent>
 
-  <name>HTTP Client Dispatcher</name>
-  <artifactId>http-client-dispatcher</artifactId>
+  <name>HTTP DI Client Sample</name>
+  <artifactId>http-client-di-sample</artifactId>
   <packaging>jar</packaging>
 
   <properties>
@@ -33,6 +33,30 @@
       <groupId>org.fido.iot</groupId>
       <artifactId>protocol</artifactId>
       <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.fido.iot</groupId>
+      <artifactId>http-client-dispatcher</artifactId>
+      <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.fido.iot</groupId>
+      <artifactId>storage-di-sample</artifactId>
+      <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-dbcp2</artifactId>
+      <version>${dbcp2.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <version>1.4.200</version>
     </dependency>
 
     <dependency>
@@ -86,9 +110,11 @@
           </execution>
         </executions>
         <configuration>
-           <executable>java</executable>
+          <executable>java</executable>
           <includeProjectDependencies>true</includeProjectDependencies>
-           <workingDirectory>${project.build.outputDirectory}</workingDirectory>
+          <workingDirectory>${project.build.outputDirectory}</workingDirectory>
+          <mainClass>org.fido.iot.web.DiClientApp</mainClass>
+          <cleanupDaemonThreads>false</cleanupDaemonThreads>
         </configuration>
       </plugin>
     </plugins>
