+++ service/protocol-samples/http-server-to0-to1-sample/pom.xml
@@ -5,13 +5,13 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <artifactId>http-dispatchers</artifactId>
+    <artifactId>protocol-samples</artifactId>
     <groupId>org.fido.iot</groupId>
     <version>0.3-SNAPSHOT</version>
   </parent>
 
-  <name>HTTP Client Dispatcher</name>
-  <artifactId>http-client-dispatcher</artifactId>
+  <name>HTTP TO0/TO1 Server Sample</name>
+  <artifactId>http-server-to0-to1-sample</artifactId>
   <packaging>jar</packaging>
 
   <properties>
@@ -20,6 +20,7 @@
     <tomcat.version>9.0.37</tomcat.version>
     <bcpkix.version>1.65</bcpkix.version>
     <dbcp2.version>2.7.0</dbcp2.version>
+    <exec.mainClass>org.fido.iot.web.RvsApp</exec.mainClass>
   </properties>
 
   <dependencies>
@@ -33,6 +34,30 @@
       <groupId>org.fido.iot</groupId>
       <artifactId>protocol</artifactId>
       <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.fido.iot</groupId>
+      <artifactId>http-server-dispatcher</artifactId>
+      <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.fido.iot</groupId>
+      <artifactId>storage-rv-sample</artifactId>
+      <version>0.3-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-dbcp2</artifactId>
+      <version>${dbcp2.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <version>1.4.200</version>
     </dependency>
 
     <dependency>
@@ -67,7 +92,7 @@
               <goal>copy-dependencies</goal>
             </goals>
             <configuration>
-              <outputDirectory>${project.build.directory}/lib/</outputDirectory>
+              <outputDirectory>target/lib</outputDirectory>
               <includeScope>runtime</includeScope>
             </configuration>
           </execution>
@@ -88,7 +113,7 @@
         <configuration>
            <executable>java</executable>
           <includeProjectDependencies>true</includeProjectDependencies>
-           <workingDirectory>${project.build.outputDirectory}</workingDirectory>
+           <workingDirectory>target/classes</workingDirectory>
         </configuration>
       </plugin>
     </plugins>
