+++ service/protocol-samples/pom.xml
@@ -10,13 +10,18 @@
     <version>0.3-SNAPSHOT</version>
   </parent>
 
-  <name>HTTP Dispatchers</name>
-  <artifactId>http-dispatchers</artifactId>
+  <name>Protocol Service Samples</name>
+  <artifactId>protocol-samples</artifactId>
   <packaging>pom</packaging>
 
   <modules>
-    <module>http-server-dispatcher</module>
-    <module>http-client-dispatcher</module>
+    <module>http-client-di-sample</module>
+    <module>http-client-to0-sample</module>
+    <module>http-client-to1-sample</module>
+    <module>http-client-to2-sample</module>
+    <module>http-server-di-sample</module>
+    <module>http-server-to0-to1-sample</module>
+    <module>http-server-to2-sample</module>
   </modules>
 
 </project>
